add Main, Start, and Home pages
