+++ client/src/pages/Main.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+const Main = () => {
+  return (
+
+  )
+};
+
+export default Main;
